services/constraint-gate/requirements.txt
flask==3.0.2
gunicorn==21.2.0
pyyaml==6.0.1
